refactor: extract FloatingKeyboardHandler from GeneralKeyboardIME (#426)
